Update: used NotPid option args and move Deprecated func to v1
