+++ robotgo_mac_win.go
@@ -16,8 +16,8 @@ package robotgo
 // GetBounds get the window bounds
 func GetBounds(pid int, args ...int) (int, int, int, int) {
 	var isPid int
-	if len(args) > 0 {
-		isPid = args[0]
+	if len(args) > 0 || NotPid {
+		isPid = 1
 	}
 
 	return internalGetBounds(pid, isPid)
@@ -26,8 +26,8 @@ func GetBounds(pid int, args ...int) (int, int, int, int) {
 // GetClient get the window client bounds
 func GetClient(pid int, args ...int) (int, int, int, int) {
 	var isPid int
-	if len(args) > 0 {
-		isPid = args[0]
+	if len(args) > 0 || NotPid {
+		isPid = 1
 	}
 
 	return internalGetClient(pid, isPid)
@@ -36,8 +36,8 @@ func GetClient(pid int, args ...int) (int, int, int, int) {
 // internalGetTitle get the window title
 func internalGetTitle(pid int, args ...int) string {
 	var isPid int
-	if len(args) > 0 {
-		isPid = args[0]
+	if len(args) > 0 || NotPid {
+		isPid = 1
 	}
 	gtitle := cgetTitle(pid, isPid)
 
@@ -54,8 +54,8 @@ func internalGetTitle(pid int, args ...int) string {
 //	robotgo.ActivePid(ids[0])
 func ActivePid(pid int, args ...int) error {
 	var isPid int
-	if len(args) > 0 {
-		isPid = args[0]
+	if len(args) > 0 || NotPid {
+		isPid = 1
 	}
 
 	internalActive(pid, isPid)
